color code update

+++ includes/cube.h
@@ -11,13 +11,13 @@
 #include <float.h>
 #include <math.h>
 #include <sys/time.h>
-// #include <X11/X.h>
-// #include <X11/keysym.h>
+#include <X11/X.h>
+#include <X11/keysym.h>
 #include "define.h"
 #include "../aux/libft.h"
 #include "../aux/libft.h"
-// #include "../mlx_linux/mlx.h"
-// #include "../mlx_linux/mlx_int.h"
+#include "../mlx_linux/mlx.h"
+#include "../mlx_linux/mlx_int.h"
 
 /* ************************************************************************** */
 /*                                   STRUCTS                                  */
@@ -93,8 +93,8 @@ typedef struct	s_map
 	char			*path;			//?? To remove?
 	char			player_direction;
 	void			*player;
-	unsigned long	C;
-	unsigned long	F;
+	int				C;
+	int				F;
 	int			map_width;
 	int			map_height;
 }				t_map;
